+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to this project will be documented in this file.
 This change log follows the conventions of
 [keepachangelog.com](http://keepachangelog.com/).
 
-## [Unreleased][unreleased]
+## [Unreleased (opus branch)][unreleased-opus]
+
+### Changed
+
+- Works with the 8.x version of Beat Link, to offer support for the Opus Quad, as long as you build and attach metadata archives for the USBs you are using, or are loading them from rekordbox or rekordbox mobile on a different computer than where BLT is running.
+
+### Added
+
+- You can now import entire playlists of tracks into shows from offline media.
+
+
+## [Unreleased (main branch)][unreleased]
 
 Nothing so far.